restructured the kernel
