+++ common/Cargo.toml
@@ -5,4 +5,4 @@ edition = "2024"
 
 [features]
 constants = []
-structures = []
+math = []
