+++ kernel/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-common = { path = "../common", features = ["structures", "constants"] }
+common = { path = "../common", features = ["constants"] }
 cpu_utils = { path = "../cpu_utils", features = [
     "paging",
     "global_descriptor_table",
