--- kernel/src/allocators/mod.rs
@@ -1,6 +1,0 @@
-use core::arch::asm;
-use cpu_utils::structures::paging::page_tables::PageTable;
-
-// pub mod free_list_allocator;
-mod bitmap;
-pub mod page_allocator;
