+++ kernel/src/memory/allocators/mod.rs
@@ -1,0 +1,2 @@
+pub mod heap_allocator;
+pub mod page_allocator;
