+++ kernel/src/memory/bitmap.rs
