+++ kernel/src/memory/mod.rs
@@ -1,0 +1,3 @@
+pub mod allocators;
+pub mod bitmap;
+pub mod memory_map;
